- Updated sql table to accommodate new data - Set order in which they displayed - Sorted ZERO k/d ratio

+++ app/src/main/java/com/ninetwozero/bf3droid/dao/PersonaStatisticsDAO.java
@@ -110,6 +110,6 @@ public class PersonaStatisticsDAO {
     }
 
     private static String kdRatio(long kills, long deaths) {
-        return NumberFormatter.format(kills / deaths);
+        return NumberFormatter.format((double)kills / deaths);
     }
 }
